CXP-1333: add product selection criteria VO
