+++ components/catalogs/back/tests/Integration/Infrastructure/Persistence/Catalog/GetCatalogQueryTest.php
@@ -6,6 +6,7 @@ namespace Akeneo\Catalogs\Test\Integration\Infrastructure\Persistence\Catalog;
 
 use Akeneo\Catalogs\Application\Exception\CatalogNotFoundException;
 use Akeneo\Catalogs\Domain\Catalog;
+use Akeneo\Catalogs\Domain\ProductSelectionCriteria;
 use Akeneo\Catalogs\Infrastructure\Persistence\Catalog\GetCatalogQuery;
 use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
 use Doctrine\DBAL\Connection;
@@ -52,13 +53,13 @@ class GetCatalogQueryTest extends IntegrationTestCase
             'Store US',
             'test',
             false,
-            [
+            new ProductSelectionCriteria([
                 [
                     'field' => 'enabled',
                     'operator' => '=',
                     'value' => true,
                 ],
-            ],
+            ]),
             ['channels' => ['ecommerce', 'print']],
             []
         );
